Actualizacion seguridad back y envio correos

+++ src/rutas/api_routes.py
@@ -45,6 +45,14 @@ def token_required(f):
         return f(current_user, *args, **kwargs)
     return decorated
 
+def check_edicion_peces(f):
+    @wraps(f)
+    def decorated_function(current_user, *args, **kwargs):
+        if not current_user.get('ed_peces'):
+            return jsonify({"success": False, "message": "Permiso denegado, no tienes permisos para editar peces"}), 403
+        return f(current_user, *args, **kwargs)
+    return decorated_function
+
 def check_admin(f):
     @wraps(f)
     def decorated_function(current_user, *args, **kwargs):
@@ -52,6 +60,102 @@ def check_admin(f):
             return jsonify({"success": False, "message": "Permiso denegado, solo para administradores"}), 403
         return f(current_user, *args, **kwargs)
     return decorated_function
+
+
+def check_edicion_ingredientes(f):
+    @wraps(f)
+    def decorated_function(current_user, *args, **kwargs):
+        if not current_user.get('ed_ingred'):
+            return jsonify({"success": False, "message": "Permiso denegado, no tienes permisos para editar ingredientes"}), 403
+        return f(current_user, *args, **kwargs)
+    return decorated_function
+
+def enviar_correo_veryfy(correo, codigo):
+    try:
+        mensaje = EmailMessage()
+        mensaje['From'] = smtp_user
+        mensaje['To'] = correo
+        mensaje['Subject'] = "Restablecer Contraseña"
+        mensaje.set_content(f"Tu código de restablecimiento es: {codigo}")
+
+        context = ssl.create_default_context()
+
+        with smtplib.SMTP_SSL(smtp_server, smtp_port, context=context) as server:
+            server.login(smtp_user, smtp_password)
+            server.send_message(mensaje)
+        
+        print("Correo enviado exitosamente")
+        return True
+    except Exception as e:
+        print(f"Error al enviar el correo: {e}")
+        return False
+    
+def enviar_correo_edicion(usuario, pez, etapa, nutriente, peso, cambio_anterior, cambio_nuevo):
+    try:
+        mensaje = EmailMessage()
+        mensaje['From'] = smtp_user
+        mensaje['To'] = "ignaciovicente04@gmail.com"
+        mensaje['Subject'] = "Actualización de Información de Peces"
+        mensaje.set_content(
+            f"El usuario {usuario} ha realizado una actualización en el pez {pez}.\n\n"
+            f"Etapa: {etapa}\n"
+            f"Nutriente: {nutriente}\n"
+            f"Peso: {peso}\n"
+            f"Se ha cambiado de {cambio_anterior} a {cambio_nuevo}"
+        )
+
+        context = ssl.create_default_context()
+
+        with smtplib.SMTP_SSL(smtp_server, smtp_port, context=context) as server:
+            server.login(smtp_user, smtp_password)
+            server.send_message(mensaje)
+        
+        print("Correo enviado exitosamente")
+        return True
+    except Exception as e:
+        print(f"Error al enviar el correo: {e}")
+        return 
+    
+
+def enviar_correo_ingrediente(accion, usuario, nombre_ingrediente, cambios=None, coste=None, stock=None):
+    try:
+        mensaje = EmailMessage()
+        mensaje['From'] = smtp_user
+        mensaje['To'] = "ignaciovicente04@gmail.com"
+
+        if accion == 'eliminar':
+            mensaje['Subject'] = "Eliminación de Ingrediente"
+            mensaje.set_content(
+                f"El usuario {usuario} ha eliminado el ingrediente {nombre_ingrediente}."
+            )
+        elif accion == 'agregar':
+            mensaje['Subject'] = "Agregación de Ingrediente"
+            mensaje.set_content(
+                f"El usuario {usuario} ha agregado un nuevo ingrediente.\n\n"
+                f"Nombre: {nombre_ingrediente}\n"
+                f"Coste: {coste}\n"
+                f"Stock: {stock}"
+            )
+        elif accion == 'editar' and cambios:
+            mensaje['Subject'] = "Edición de Ingrediente"
+            cambios_str = "\n".join([f"{campo}: de {anterior} a {nuevo}" for campo, (anterior, nuevo) in cambios.items()])
+            mensaje.set_content(
+                f"El usuario {usuario} ha editado el ingrediente {nombre_ingrediente}.\n\n"
+                f"Cambios realizados:\n{cambios_str}"
+            )
+
+        context = ssl.create_default_context()
+
+        with smtplib.SMTP_SSL(smtp_server, smtp_port, context=context) as server:
+            server.login(smtp_user, smtp_password)
+            server.send_message(mensaje)
+        
+        print("Correo enviado exitosamente")
+        return True
+    except Exception as e:
+        print(f"Error al enviar el correo: {e}")
+        return False
+
 
 # Ruta para registrar un nuevo usuario
 @api.route('/register', methods=['POST'])
@@ -66,14 +170,19 @@ def register_user():
     hashed_password = generate_password_hash(data['password'], method='pbkdf2:sha256')
     
     # Verificar si el correo es el del administrador
-    is_admin = data['correo'] == 'katherine.alveal@aquapacifico.cl' or 'miguelpradenas@hotmail.com'  # Cambia esto al correo del administrador real
+    is_admin = data['correo'] in ['katherine.alveal@aquapacifico.cl', 'miguelpradenas@hotmail.com']  # Cambia esto al correo del administrador real
     
+    # Establecer permisos de edición en True si es administrador
+    ed_peces = ed_ingred = is_admin
+
     result = db.usuarios.insert_one({
         "nombre": data.get('nombre'),
         "correo": data.get('correo'),
         "cargo": data.get('cargo'),
         "password": hashed_password,
-        "is_admin": is_admin
+        "is_admin": is_admin,
+        "ed_peces": ed_peces,
+        "ed_ingred": ed_ingred
     })
     return jsonify({"success": True, "inserted_id": str(result.inserted_id)})
 
@@ -101,7 +210,18 @@ def login():
 @api.route('/profile', methods=['GET'])
 @token_required
 def profile(current_user):
-    return jsonify({"success": True, "user": {"nombre": current_user['nombre'], "correo": current_user['correo'], "cargo": current_user['cargo'], "is_admin": current_user.get('is_admin', False)}})
+    return jsonify({
+        "success": True, 
+        "user": {
+            "nombre": current_user['nombre'],
+            "correo": current_user['correo'],
+            "cargo": current_user['cargo'],
+            "is_admin": current_user.get('is_admin', False),
+            "ed_peces": current_user.get('ed_peces', False),
+            "ed_ingred": current_user.get('ed_ingred', False)
+        }
+    })
+
 
 
 # Rutas de cargos
@@ -117,25 +237,6 @@ def add_cargo():
     result = db.cargos.insert_one({"nombre": data["nombre"]})
     return jsonify({"inserted_id": str(result.inserted_id)})
 
-def enviar_correo(correo, codigo):
-    try:
-        mensaje = EmailMessage()
-        mensaje['From'] = smtp_user
-        mensaje['To'] = correo
-        mensaje['Subject'] = "Restablecer Contraseña"
-        mensaje.set_content(f"Tu código de restablecimiento es: {codigo}")
-
-        context = ssl.create_default_context()
-
-        with smtplib.SMTP_SSL(smtp_server, smtp_port, context=context) as server:
-            server.login(smtp_user, smtp_password)
-            server.send_message(mensaje)
-        
-        print("Correo enviado exitosamente")
-        return True
-    except Exception as e:
-        print(f"Error al enviar el correo: {e}")
-        return False
 
 @api.route('/forgot-password', methods=['POST'])
 def forgot_password():
@@ -150,7 +251,7 @@ def forgot_password():
     if user:
         codigo = ''.join(random.choices(string.ascii_uppercase + string.digits, k=6))
         reset_codes[email] = codigo
-        if enviar_correo(email, codigo):
+        if enviar_correo_veryfy(email, codigo):
             print("Correo enviado exitosamente")
             return jsonify({"success": True}), 200
         else:
@@ -390,7 +491,7 @@ def corvina_food():
 
 @api.route('/congrio-edit', methods=['GET'])
 @token_required
-@check_admin
+@check_edicion_peces
 def get_congrio_edit_data(current_user):
     try:
         congrio_data = list(db.peces.find({"nombre": "Congrio"}))
@@ -403,7 +504,7 @@ def get_congrio_edit_data(current_user):
 
 @api.route('/pez/<id>', methods=['GET', 'PUT'])
 @token_required
-@check_admin
+@check_edicion_peces
 def edit_pez(current_user, id):
     try:
         if request.method == 'GET':
@@ -417,12 +518,51 @@ def edit_pez(current_user, id):
             data = request.json
             data.pop('_id', None)  # Evitar modificar el _id
 
-            convert_and_process(data)  # Procesa el diccionario completo
+            # Obtener el valor anterior
+            pez_anterior = db.peces.find_one({"_id": ObjectId(id)})
+            if not pez_anterior:
+                return jsonify({"success": False, "message": "No se encontró el pez a actualizar"}), 404
 
+            convert_and_process(data)
+
+            # Realizar la actualización
             result = db.peces.update_one({"_id": ObjectId(id)}, {"$set": data})
 
             if result.matched_count == 0:
                 return jsonify({"success": False, "message": "No se encontró el pez a actualizar"}), 404
+
+            # Obtener los valores del cambio
+            pez_actual = db.peces.find_one({"_id": ObjectId(id)})
+            
+            # Extraer el nutriente y el peso específico que se cambió
+            nutriente_cambiado = None
+            peso_cambiado = None
+            cambio_anterior = None
+            cambio_nuevo = None
+
+            for peso, requerimientos in data.get('requerimientos', {}).items():
+                for nutriente, valores in requerimientos.items():
+                    if (pez_anterior['requerimientos'][peso][nutriente]['min'] != pez_actual['requerimientos'][peso][nutriente]['min'] or
+                        pez_anterior['requerimientos'][peso][nutriente]['max'] != pez_actual['requerimientos'][peso][nutriente]['max']):
+                        nutriente_cambiado = nutriente
+                        peso_cambiado = peso
+                        cambio_anterior = pez_anterior['requerimientos'][peso][nutriente]
+                        cambio_nuevo = pez_actual['requerimientos'][peso][nutriente]
+                        break
+                if nutriente_cambiado:
+                    break
+            
+            if nutriente_cambiado and peso_cambiado:
+                enviar_correo_edicion(
+                    current_user['nombre'],
+                    pez_actual['nombre'],
+                    pez_actual['etapa'],
+                    nutriente_cambiado,
+                    peso_cambiado,
+                    f"min: {cambio_anterior['min']}, max: {cambio_anterior['max']}",
+                    f"min: {cambio_nuevo['min']}, max: {cambio_nuevo['max']}"
+                )
+            
             print(f"Datos después de la actualización: {db.peces.find_one({'_id': ObjectId(id)})}")
 
             return jsonify({"success": True, "message": "Datos actualizados correctamente"}), 200
@@ -434,7 +574,7 @@ def edit_pez(current_user, id):
 
 @api.route('/cojinova-edit', methods=['GET'])
 @token_required
-@check_admin
+@check_edicion_peces
 def get_cojinova_edit_data(current_user):
     try:
         cojinova_data = list(db.peces.find({"nombre": "Cojinova"}))
@@ -446,7 +586,7 @@ def get_cojinova_edit_data(current_user):
     
 @api.route('/corvina-edit', methods=['GET'])
 @token_required
-@check_admin
+@check_edicion_peces
 def get_corvina_edit_data(current_user):
     try:
         corvina_data = list(db.peces.find({"nombre": "Corvina"}))
@@ -458,7 +598,7 @@ def get_corvina_edit_data(current_user):
 
 @api.route('/palometa-edit', methods=['GET'])
 @token_required
-@check_admin
+@check_edicion_peces
 def get_palometa_edit_data(current_user):
     try:
         palometa_data = list(db.peces.find({"nombre": "Palometa"}))
@@ -480,7 +620,7 @@ def get_ingredientes():
 
 @api.route('/ingredientes', methods=['POST'])
 @token_required
-@check_admin
+@check_edicion_ingredientes
 def add_ingrediente(current_user):
     data = request.json
     try:
@@ -493,16 +633,31 @@ def add_ingrediente(current_user):
             "stock": float(data.get("stock"))
         }
         db.ingredientes.insert_one(nuevo_ingrediente)
+
+        # Enviar correo de notificación para la agregación
+        enviar_correo_ingrediente(
+            accion='agregar',
+            usuario=current_user['nombre'],
+            nombre_ingrediente=nuevo_ingrediente["nombre"],
+            coste=nuevo_ingrediente["coste"],
+            stock=nuevo_ingrediente["stock"]
+        )
+
         return jsonify({"success": True, "message": "Ingrediente agregado correctamente"}), 201
     except Exception as e:
         return jsonify({"success": False, "message": str(e)}), 500
 
 @api.route('/ingredientes/<id>', methods=['PUT'])
 @token_required
-@check_admin
+@check_edicion_ingredientes
 def update_ingrediente(current_user, id):
     data = request.json
     try:
+        # Obtener el valor anterior
+        ingrediente_anterior = db.ingredientes.find_one({"_id": ObjectId(id)})
+        if not ingrediente_anterior:
+            return jsonify({"success": False, "message": "Ingrediente no encontrado"}), 404
+
         ingrediente_actualizado = {
             "nombre": data.get("nombre"),
             "coste": float(data.get("coste")),
@@ -512,16 +667,96 @@ def update_ingrediente(current_user, id):
             "stock": float(data.get("stock"))
         }
         db.ingredientes.update_one({"_id": ObjectId(id)}, {"$set": ingrediente_actualizado})
+
+        # Identificar los campos que cambiaron
+        cambios = {}
+        for campo in ingrediente_actualizado:
+            if ingrediente_actualizado[campo] != ingrediente_anterior[campo]:
+                cambios[campo] = (ingrediente_anterior[campo], ingrediente_actualizado[campo])
+
+        # Enviar correo de notificación para la edición
+        correo_enviado = enviar_correo_ingrediente(
+            accion='editar',
+            usuario=current_user['nombre'],
+            nombre_ingrediente=ingrediente_actualizado["nombre"],
+            cambios=cambios
+        )
+
+        if not correo_enviado:
+            print("⚠️ Error al enviar el correo de notificación")
+            return jsonify({"success": False, "message": "Error al enviar el correo de notificación"}), 500
+
         return jsonify({"success": True, "message": "Ingrediente actualizado correctamente"}), 200
     except Exception as e:
+        print(f"⚠️ Error en PUT /ingredientes/{id}: {e}")  # Mostrar error en la terminal
         return jsonify({"success": False, "message": str(e)}), 500
+
 
 @api.route('/ingredientes/<id>', methods=['DELETE'])
 @token_required
-@check_admin
+@check_edicion_ingredientes
 def delete_ingrediente(current_user, id):
     try:
+        # Obtener el ingrediente antes de eliminarlo
+        ingrediente = db.ingredientes.find_one({"_id": ObjectId(id)})
+        if not ingrediente:
+            return jsonify({"success": False, "message": "Ingrediente no encontrado"}), 404
+        
         db.ingredientes.delete_one({"_id": ObjectId(id)})
+
+        # Enviar correo de notificación para la eliminación
+        enviar_correo_ingrediente(
+            accion='eliminar',
+            usuario=current_user['nombre'],
+            nombre_ingrediente=ingrediente["nombre"]
+        )
+
         return jsonify({"success": True, "message": "Ingrediente eliminado correctamente"}), 200
+    except Exception as e:
+        return jsonify({"success": False, "message": str(e)}), 500
+
+
+@api.route('/usuarios', methods=['GET'])
+@token_required
+@check_admin
+def get_usuarios(current_user):
+    try:
+        # Excluir la contraseña de la respuesta
+        usuarios = list(db.usuarios.find({}, {"password": 0}))
+        for usuario in usuarios:
+            usuario['_id'] = str(usuario['_id'])
+        return jsonify(usuarios), 200
+    except Exception as e:
+        return jsonify({"success": False, "message": str(e)}), 500
+
+@api.route('/usuarios/<id>', methods=['PUT'])
+@token_required
+@check_admin
+def update_usuario(current_user, id):
+    data = request.json
+    try:
+        campo = data.get('campo')
+        if campo not in ['ed_peces', 'ed_ingred']:
+            return jsonify({"success": False, "message": "Campo no válido"}), 400
+
+        usuario = db.usuarios.find_one({"_id": ObjectId(id)})
+        if not usuario:
+            return jsonify({"success": False, "message": "Usuario no encontrado"}), 404
+
+        # Actualizar el campo correspondiente
+        db.usuarios.update_one({"_id": ObjectId(id)}, {"$set": {campo: not usuario[campo]}})
+        return jsonify({"success": True, "message": "Permiso actualizado correctamente"}), 200
+    except Exception as e:
+        return jsonify({"success": False, "message": str(e)}), 500
+
+@api.route('/usuarios/<id>', methods=['DELETE'])
+@token_required
+@check_admin  # O el decorador que uses para verificar permisos de eliminación
+def delete_usuario(current_user, id):
+    try:
+        result = db.usuarios.delete_one({"_id": ObjectId(id)})
+        if result.deleted_count == 0:
+            return jsonify({"success": False, "message": "Usuario no encontrado"}), 404
+        return jsonify({"success": True, "message": "Usuario eliminado correctamente"}), 200
     except Exception as e:
         return jsonify({"success": False, "message": str(e)}), 500
